[FEAT] canvas socket 세팅 및 block socket url 변경
